// search/public/tinymce/tinymce.d.ts
interface EventDispatcherConstructor<T extends {}> {
  readonly prototype: EventDispatcher<T>;
  new (settings?: EventDispatcherSettings): EventDispatcher<T>;
  isNative: (name: string) => boolean;
}
declare class EventDispatcher<T extends {}> {
  static isNative(name: string): boolean;
  private readonly settings;
  private readonly scope;
  private readonly toggleEvent;
  private bindings;
  constructor(settings?: EventDispatcherSettings);
  fire<K extends string, U extends MappedEvent<T, K>>(
    name: K,
    args?: U,
  ): EditorEvent<U>;
  dispatch<K extends string, U extends MappedEvent<T, K>>(
    name: K,
    args?: U,
  ): EditorEvent<U>;
  on<K extends string>(
    name: K,
    callback:
      | false
      | ((event: EditorEvent<MappedEvent<T, K>>) => void | boolean),
    prepend?: boolean,
    extra?: {},
  ): this;
  off<K extends string>(
    name?: K,
    callback?: (event: EditorEvent<MappedEvent<T, K>>) => void,
  ): this;
  once<K extends string>(
    name: K,
    callback: (event: EditorEvent<MappedEvent<T, K>>) => void,
    prepend?: boolean,
  ): this;
  has(name: string): boolean;
}
// ...
interface Rect {
  inflate: (rect: GeomRect, w: number, h: number) => GeomRect;
  relativePosition: (
    rect: GeomRect,
    targetRect: GeomRect,
    rel: string,
  ) => GeomRect;
  findBestRelativePosition: (
    rect: GeomRect,
    targetRect: GeomRect,
    constrainRect: GeomRect,
    rels: string[],
  ) => string | null;
  intersect: (rect: GeomRect, cropRect: GeomRect) => GeomRect | null;
  clamp: (rect: GeomRect, clampRect: GeomRect, fixedSize?: boolean) => GeomRect;
  create: (x: number, y: number, w: number, h: number) => GeomRect;
  fromClientRect: (clientRect: DOMRect) => GeomRect;
}
// ...
interface EventUtilsConstructor {
  readonly prototype: EventUtils;
  new (): EventUtils;
  Event: EventUtils;
}
declare class EventUtils {
  static Event: EventUtils;
  domLoaded: boolean;
  events: Record<number, Record<string, CallbackList<any>>>;
  private readonly expando;
  private hasFocusIn;
  private count;
  constructor();
  bind<K extends keyof HTMLElementEventMap>(
    target: any,
    name: K,
    callback: EventUtilsCallback<HTMLElementEventMap[K]>,
    scope?: any,
  ): EventUtilsCallback<HTMLElementEventMap[K]>;
  bind<T = any>(
    target: any,
    names: string,
    callback: EventUtilsCallback<T>,
    scope?: any,
  ): EventUtilsCallback<T>;
  unbind<K extends keyof HTMLElementEventMap>(
    target: any,
    name: K,
    callback?: EventUtilsCallback<HTMLElementEventMap[K]>,
  ): this;
  unbind<T = any>(
    target: any,
    names: string,
    callback?: EventUtilsCallback<T>,
  ): this;
  unbind(target: any): this;
  fire(target: any, name: string, args?: {}): this;
  dispatch(target: any, name: string, args?: {}): this;
  clean(target: any): this;
  destroy(): void;
  cancel<T>(e: EventUtilsEvent<T>): boolean;
  private executeHandlers;
}
// ...
declare type TranslatedString = string;
interface I18n {
  getData: () => Record<string, Record<string, string>>;
  setCode: (newCode: string) => void;
  getCode: () => string;
  add: (code: string, items: Record<string, string>) => void;
  translate: (text: Untranslated) => TranslatedString;
  isRtl: () => boolean;
  hasCode: (code: string) => boolean;
}
interface Observable<T extends {}> {
  fire<K extends string, U extends MappedEvent<T, K>>(
    name: K,
    args?: U,
    bubble?: boolean,
  ): EditorEvent<U>;
  dispatch<K extends string, U extends MappedEvent<T, K>>(
    name: K,
    args?: U,
    bubble?: boolean,
  ): EditorEvent<U>;
  on<K extends string>(
    name: K,
    callback: (event: EditorEvent<MappedEvent<T, K>>) => void,
    prepend?: boolean,
  ): EventDispatcher<T>;
  off<K extends string>(
    name?: K,
    callback?: (event: EditorEvent<MappedEvent<T, K>>) => void,
  ): EventDispatcher<T>;
  once<K extends string>(
    name: K,
    callback: (event: EditorEvent<MappedEvent<T, K>>) => void,
  ): EventDispatcher<T>;
  hasEventListeners(name: string): boolean;
}
interface URISettings {
  base_uri?: URI;
}
interface URIConstructor {
  readonly prototype: URI;
  new (url: string, settings?: URISettings): URI;
  getDocumentBaseUrl: (loc: {
    protocol: string;
    host?: string;
    href?: string;
    pathname?: string;
  }) => string;
  parseDataUri: (uri: string) => {
    type: string;
    data: string;
  };
}
interface SafeUriOptions {
  readonly allow_html_data_urls?: boolean;
  readonly allow_script_urls?: boolean;
  readonly allow_svg_data_urls?: boolean;
}
declare class URI {
  static parseDataUri(uri: string): {
    type: string | undefined;
    data: string;
  };
  static isDomSafe(
    uri: string,
    context?: string,
    options?: SafeUriOptions,
  ): boolean;
  static getDocumentBaseUrl(loc: {
    protocol: string;
    host?: string;
    href?: string;
    pathname?: string;
  }): string;
  source: string;
  protocol: string | undefined;
  authority: string | undefined;
  userInfo: string | undefined;
  user: string | undefined;
  password: string | undefined;
  host: string | undefined;
  port: string | undefined;
  relative: string | undefined;
  path: string;
  directory: string;
  file: string | undefined;
  query: string | undefined;
  anchor: string | undefined;
  settings: URISettings;
  constructor(url: string, settings?: URISettings);
  setPath(path: string): void;
  toRelative(uri: string): string;
  toAbsolute(uri: string, noHost?: boolean): string;
  isSameOrigin(uri: URI): boolean;
  toRelPath(base: string, path: string): string;
  toAbsPath(base: string, path: string): string;
  getURI(noProtoHost?: boolean): string;
}
interface EditorManager extends Observable<EditorManagerEventMap> {
  defaultOptions: RawEditorOptions;
  majorVersion: string;
  minorVersion: string;
  releaseDate: string;
  activeEditor: Editor | null;
  focusedEditor: Editor | null;
  baseURI: URI;
  baseURL: string;
  documentBaseURL: string;
  i18n: I18n;
  suffix: string;
  add(this: EditorManager, editor: Editor): Editor;
  addI18n: (code: string, item: Record<string, string>) => void;
  createEditor(
    this: EditorManager,
    id: string,
    options: RawEditorOptions,
  ): Editor;
  execCommand(
    this: EditorManager,
    cmd: string,
    ui: boolean,
    value: any,
  ): boolean;
  get(this: EditorManager): Editor[];
  get(this: EditorManager, id: number | string): Editor | null;
  init(this: EditorManager, options: RawEditorOptions): Promise<Editor[]>;
  overrideDefaults(
    this: EditorManager,
    defaultOptions: Partial<RawEditorOptions>,
  ): void;
  remove(this: EditorManager): void;
  remove(this: EditorManager, selector: string): void;
  remove(this: EditorManager, editor: Editor): Editor | null;
  setActive(this: EditorManager, editor: Editor): void;
  setup(this: EditorManager): void;
  translate: (text: Untranslated) => TranslatedString;
  triggerSave: () => void;
  _setBaseUrl(this: EditorManager, baseUrl: string): void;
}
// ...
interface DomTreeWalkerConstructor {
  readonly prototype: DomTreeWalker;
  new (startNode: Node, rootNode: Node): DomTreeWalker;
}
// ...
interface Delay {
  setEditorInterval: (
    editor: Editor,
    callback: () => void,
    time?: number,
  ) => number;
  setEditorTimeout: (
    editor: Editor,
    callback: () => void,
    time?: number,
  ) => number;
}
declare type UploadResult = UploadResult$2;
interface ImageUploader {
  upload: (
    blobInfos: BlobInfo[],
    showNotification?: boolean,
  ) => Promise<UploadResult[]>;
}
// ...
interface Tools {
  is: (obj: any, type?: string) => boolean;
  isArray: <T>(arr: any) => arr is Array<T>;
  inArray: <T>(arr: ArrayLike<T>, value: T) => number;
  grep: {
    <T>(
      arr: ArrayLike<T> | null | undefined,
      pred?: ArrayCallback<T, boolean>,
    ): T[];
    <T>(
      arr: Record<string, T> | null | undefined,
      pred?: ObjCallback<T, boolean>,
    ): T[];
  };
  trim: (str: string | null | undefined) => string;
  toArray: <T>(obj: ArrayLike<T>) => T[];
  hasOwn: (obj: any, name: string) => boolean;
  makeMap: (
    items: ArrayLike<string> | string | undefined,
    delim?: string | RegExp,
    map?: Record<string, {}>,
  ) => Record<string, {}>;
  each: {
    <T>(
      arr: ArrayLike<T> | null | undefined,
      cb: ArrayCallback<T, void | boolean>,
      scope?: any,
    ): boolean;
    <T>(
      obj: Record<string, T> | null | undefined,
      cb: ObjCallback<T, void | boolean>,
      scope?: any,
    ): boolean;
  };
  map: {
    <T, R>(arr: ArrayLike<T> | null | undefined, cb: ArrayCallback<T, R>): R[];
    <T, R>(
      obj: Record<string, T> | null | undefined,
      cb: ObjCallback<T, R>,
    ): R[];
  };
  extend: (obj: Object, ext: Object, ...objs: Object[]) => any;
  walk: <T extends Record<string, any>>(
    obj: T,
    f: WalkCallback<T>,
    n?: keyof T,
    scope?: any,
  ) => void;
  resolve: (path: string, o?: Object) => any;
  explode: (s: string | string[], d?: string | RegExp) => string[];
  _addCacheSuffix: (url: string) => string;
}
interface KeyboardLikeEvent {
  shiftKey: boolean;
  ctrlKey: boolean;
  altKey: boolean;
  metaKey: boolean;
}
interface VK {
  BACKSPACE: number;
  DELETE: number;
  DOWN: number;
  ENTER: number;
  ESC: number;
  LEFT: number;
  RIGHT: number;
  SPACEBAR: number;
  TAB: number;
  UP: number;
  PAGE_UP: number;
  PAGE_DOWN: number;
  END: number;
  HOME: number;
  modifierPressed: (e: KeyboardLikeEvent) => boolean;
  metaKeyPressed: (e: KeyboardLikeEvent) => boolean;
}
interface DOMUtilsNamespace {
  (doc: Document, settings: Partial<DOMUtilsSettings>): DOMUtils;
  DOM: DOMUtils;
  nodeIndex: (node: Node, normalized?: boolean) => number;
}
// ...
interface TinyMCE extends EditorManager {
  geom: {
    Rect: Rect;
  };
  util: {
    Delay: Delay;
    Tools: Tools;
    VK: VK;
    URI: URIConstructor;
    EventDispatcher: EventDispatcherConstructor<any>;
    Observable: Observable<any>;
    I18n: I18n;
    LocalStorage: Storage;
    ImageUploader: ImageUploader;
  };
  dom: {
    EventUtils: EventUtilsConstructor;
    TreeWalker: DomTreeWalkerConstructor;
    TextSeeker: (
      dom: DOMUtils,
      isBlockBoundary?: (node: Node) => boolean,
    ) => TextSeeker;
    DOMUtils: DOMUtilsNamespace;
    ScriptLoader: ScriptLoaderConstructor;
    RangeUtils: RangeUtilsNamespace;
    Serializer: (
      settings: DomSerializerSettings,
      editor?: Editor,
    ) => DomSerializer;
    ControlSelection: (
      selection: EditorSelection,
      editor: Editor,
    ) => ControlSelection;
    BookmarkManager: BookmarkManagerNamespace;
    Selection: (
      dom: DOMUtils,
      win: Window,
      serializer: DomSerializer,
      editor: Editor,
    ) => EditorSelection;
    StyleSheetLoader: (
      documentOrShadowRoot: Document | ShadowRoot,
      settings: StyleSheetLoaderSettings,
    ) => StyleSheetLoader;
    Event: EventUtils;
  };
  html: {
    Styles: (settings?: StylesSettings, schema?: Schema) => Styles;
    Entities: Entities;
    Node: AstNodeConstructor;
    Schema: (settings?: SchemaSettings) => Schema;
    DomParser: (settings?: DomParserSettings, schema?: Schema) => DomParser;
    Writer: (settings?: WriterSettings) => Writer;
    Serializer: (
      settings?: HtmlSerializerSettings,
      schema?: Schema,
    ) => HtmlSerializer;
  };
  AddOnManager: AddOnManagerNamespace;
  Annotator: (editor: Editor) => Annotator;
  Editor: EditorConstructor;
  EditorCommands: EditorCommandsConstructor;
  EditorManager: EditorManager;
  EditorObservable: EditorObservable;
  Env: Env;
  FocusManager: FocusManager;
  Formatter: (editor: Editor) => Formatter;
  NotificationManager: (editor: Editor) => NotificationManager;
  Shortcuts: ShortcutsConstructor;
  UndoManager: (editor: Editor) => UndoManager;
  WindowManager: (editor: Editor) => WindowManager;
  DOM: DOMUtils;
  ScriptLoader: ScriptLoader;
  PluginManager: PluginManager;
  ThemeManager: ThemeManager;
  ModelManager: ModelManager;
  IconManager: IconManager;
  Resource: Resource;
  FakeClipboard: FakeClipboard;
  trim: Tools["trim"];
  isArray: Tools["isArray"];
  is: Tools["is"];
  toArray: Tools["toArray"];
  makeMap: Tools["makeMap"];
  each: Tools["each"];
  map: Tools["map"];
  grep: Tools["grep"];
  inArray: Tools["inArray"];
  extend: Tools["extend"];
  walk: Tools["walk"];
  resolve: Tools["resolve"];
  explode: Tools["explode"];
  _addCacheSuffix: Tools["_addCacheSuffix"];
}